Create a separate crate for workunit_store
